requirements.txt
Django>=4.0
djangorestframework
mysqlclient
gunicorn
whitenoise
psycopg2-binary
drf-yasg
